Corrected RotaryEncoder argument type mismatch (int->long) Memory spare: - Button poll interval moved to global (not bound to single button) - Handler callbacks for MFButton/MFEncoder moved to class static

+++ libraries/RotaryEncoder/RotaryEncoder.cpp
@@ -52,7 +52,7 @@ void RotaryEncoder::initialize(int pin1, int pin2) {
 } // RotaryEncoder()
 
 
-void RotaryEncoder::setMinMax(int min, int max) {
+void RotaryEncoder::setMinMax(long min, long max) {
 	_minValue = min;
 	_maxValue = max;
 }
